Fix real voice input pipeline for first Windows test
- RealHotkeyListener now actually implements hold-to-talk: it hooks raw
  key events and tracks the hotkey's own keys, firing on_press once all
  are down and on_release the moment any comes back up. The previous
  keyboard.add_hotkey() call only ever fired once and never called
  on_release, so recording could start but would never stop.
- RealAudioRecorder now records int16 PCM explicitly. sounddevice
  defaults to float32; writing that through save_wav's 16-bit WAV header
  produced a corrupt/garbled file.
- FasterWhisperEngine now accepts download_root, and the ASR factory
  points it at the same data_dir/models/<size> path
  scripts/download_asr_model.py downloads into -- previously the engine
  ignored that and looked in the default HF cache, so a downloaded model
  was never actually used at runtime.
- main.py voice loop: replaced the busy `while True: pass` with a sleep,
  and one bad command no longer kills the whole session.

+++ orbit/voice/recorder.py
@@ -61,8 +61,11 @@ class RealAudioRecorder(AudioRecorder):
 
     def start(self) -> None:
         self._frames = []
+        # int16 explicitly -- `save_wav` writes 16-bit PCM (sampwidth=2).
+        # sounddevice defaults to float32, which would silently produce a
+        # corrupt/garbled WAV file if written out with a 16-bit header.
         self._stream = self._sd.InputStream(
-            samplerate=self.samplerate, channels=self.channels, callback=self._callback
+            samplerate=self.samplerate, channels=self.channels, dtype="int16", callback=self._callback
         )
         self._stream.start()
